requirements.txt
filetype
gradio
gradio-rangeslider
numpy
onnx
onnxruntime
opencv-python
psutil
pydantic
tqdm
scipy
opencv-python
google-cloud-storage
